Implement Chapter 10 - Using Fragment Arguments
When we go from the list to detail view, we pass the crimeId so the detail view can show the
appropriate crime. We also update the RecyclerView onResume so that it can update based
on the changed data. However, that update means that it's changing *all* of the fields even
though only one has changed. Challenge (not yet implemented) is to do a micro update.