Implement Chapter 13 - The Toolbar
Implemented a menu bar with just one item that allows for adding new items.
As a result, the list of random crimes is no longer needed. Had to add back
`onResume` in order to get the newly created item to appear. Need to investigate
why it didn't.

Also added a custom menu add icon generated from the system + icon